Modif compatibilité Firefox

+++ view/backend/adminManager.php
@@ -27,12 +27,9 @@ if (isset($_SESSION['login']) && ($_SESSION['role'] == 'Administrateur')) {
                             <div>image associée : <a href="public/uploads/<?= $post->postImg() ; ?>" target="_blank"><?= $post->postImg() ; ?></a></div>
                         <?php } ?>
                         <div class="btnGestion">
-                            <button type="button" class="btn btn-success"><a
-                                        href="index.php?action=editPost&&id=<?= $post->id() ?>"><i style="font-size:18px" class="fa">&#xf0a4;</i> Editer</a>
+                            <a href="index.php?action=editPost&&id=<?= $post->id() ?>" class="btn btn-success"><i style="font-size:18px" class="fa">&#xf0a4;</i> Editer</a>
                             </button>
-                            <button type="button" class="btn btn-danger"><a
-                                        href="index.php?action=deletePost&&id=<?= $post->id() ?>"
-                                        onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article ?');"><i style="font-size:18px" class="fa">&#xf088;</i> Supprimer</a>
+                            <a href="index.php?action=deletePost&&id=<?= $post->id() ?>" class="btn btn-danger" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article ?');"><i style="font-size:18px" class="fa">&#xf088;</i> Supprimer</a>
                             </button>
                         </div>
                     <?php }; ?>
@@ -41,8 +38,9 @@ if (isset($_SESSION['login']) && ($_SESSION['role'] == 'Administrateur')) {
         </div>
     </div>
 <?php } else {
-    echo('<center>Vous n\'êtes pas autorisé à accéder à cette partie du site</center>');
     header("Refresh: 2; URL=index.php");
+    echo('<center>Vous n\'êtes pas autorisé à accéder à cette partie du site</center>');
+    exit;
 }
 $content = ob_get_clean();
 $title = "Jean Forteroche - Administration";
